backend/cftv/serializers.py
from rest_framework import serializers
from .models import Server, Camera

class CameraSerializer(serializers.ModelSerializer):
    class Meta:
        model = Camera
        fields = ('id', 'memory_name', 'configuration')

class ServerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Server
        fields = ('id', 'ip', 'port')