backend/requirements.txt
asgiref
Django==3.1
pytz==2020.1
sqlparse==0.3.1
djangorestframework
dj-rest-auth
django-allauth
djangorestframework-simplejwt
markdown
django-import-export
django-filter
opencv-python